Move Controller to controller package
