+++ src/main/java/com/codecool/queststore/App.java
@@ -1,5 +1,7 @@
 package com.codecool.queststore;
 
+import com.codecool.queststore.controller.Controller;
+
 public class App {
     public static void main(String[] args) {
         Controller controller = new Controller();
